REPO-3259 - Resolve cross dependency between AuthorityService and AuthorityDAOImpl
