+++ src/main/resources/alfresco/authority-services-context.xml
@@ -77,9 +77,6 @@
         <property name="personService">
             <ref bean="personService" />
         </property>
-        <property name="authorityService">
-            <ref bean="authorityService" />
-        </property>
         <property name="tenantService">
             <ref bean="tenantService" />
         </property>
@@ -109,7 +106,16 @@
         <property name="authorityBridgeTableCache" ref="authorityBridgeTableCache" />
         <property name="useBridgeTable" value="${authority.useBridgeTable}" />
     </bean>
-    
+
+    <bean id="authorityTypeBehaviour" class="org.alfresco.repo.security.authority.AuthorityTypeBehaviour" init-method="init">
+        <property name="authorityService">
+            <ref bean="authorityService" />
+        </property>
+        <property name="policyComponent">
+            <ref bean="policyComponent"/>
+        </property>
+    </bean>
+
     <bean name="authorityBridgeDAO" class="org.alfresco.repo.security.authority.AuthorityBridgeDAOImpl">
         <property name="tenantService" ref="tenantService"/>
         <property name="nodeDAO" ref="nodeDAO"/>
